move extension types to `types` package
