+++ library/react-connect-kit/src/ExtensionsProvider/types.ts
@@ -1,8 +1,7 @@
 /* @license Copyright 2024 w3ux authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
-import type { ExtensionAccount, ExtensionStatus, VoidFn } from '@w3ux/types'
-import type { FunctionComponent, SVGProps } from 'react'
+import type { ExtensionStatus } from '@w3ux/types'
 
 // Extensions context interface.
 export interface ExtensionsContextInterface {
@@ -13,31 +12,4 @@ export interface ExtensionsContextInterface {
   extensionInstalled: (id: string) => boolean
   extensionCanConnect: (id: string) => boolean
   extensionHasFeature: (id: string, feature: string) => boolean
-}
-
-// Top level required properties the extension must expose via their `injectedWeb3` entry.
-export interface ExtensionInjected extends ExtensionConfig {
-  id: string
-  enable: (n: string) => Promise<ExtensionInterface>
-}
-
-// Required properties `enable` must provide after resolution.
-export interface ExtensionInterface {
-  accounts: {
-    subscribe: (a: (b: ExtensionAccount[]) => void) => VoidFn
-    get: () => Promise<ExtensionAccount[]>
-  }
-  provider: unknown
-  metadata: unknown
-  signer: unknown
-}
-
-// Configuration item of an extension.
-export interface ExtensionConfig {
-  id: string
-  title: string
-  icon: FunctionComponent<
-    SVGProps<SVGSVGElement> & { title?: string | undefined }
-  >
-  url: string
 }
