+++ library/react-connect-kit/src/types.ts
@@ -1,10 +1,7 @@
 /* @license Copyright 2024 w3ux authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
-import type {
-  ExtensionInjected,
-  ExtensionInterface,
-} from './ExtensionsProvider/types'
+import type { ExtensionInjected, ExtensionInterface } from '@w3ux/types'
 
 declare global {
   interface Window {
