+++ library/types/src/accounts.ts
@@ -1,19 +1,12 @@
 /* @license Copyright 2024 w3ux authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
+import type { ExtensionAccount } from './extensions'
+
 export type AccountSource = 'extension' | 'external' | 'ledger' | 'vault'
 
-export type ExternalAccountAddedBy = 'system' | 'user'
+export type AccountAddedBy = 'system' | 'user'
 
-export type ExtensionStatus = 'installed' | 'not_authenticated' | 'connected'
-
-// Miscellaneous metadata added to an extension.
-export interface ExtensionMetadata {
-  addedBy?: ExternalAccountAddedBy
-  source: string
-}
-
-// General account type that can be imported.
 export type ImportedAccount =
   | ExtensionAccount
   | ExternalAccount
@@ -21,18 +14,12 @@ export type ImportedAccount =
   | VaultAccount
   | WCAccount
 
-export interface ExtensionAccount extends ExtensionMetadata {
-  address: string
-  name: string
-  signer?: unknown
-}
-
 export interface ExternalAccount {
   address: string
   network: string
   name: string
   source: string
-  addedBy: ExternalAccountAddedBy
+  addedBy: AccountAddedBy
 }
 
 export interface LedgerAccount {
