+++ library/types/src/index.ts
@@ -3,4 +3,5 @@ SPDX-License-Identifier: GPL-3.0-only */
 
 export * from './accounts'
 export * from './common'
+export * from './extensions'
 export * from './time'
