+++ library/utils/package.json
@@ -13,7 +13,7 @@
   },
   "devDependencies": {
     "@types/react": "^19.0.12",
-    "@w3ux/types": "^2.0.3",
+    "@w3ux/types": "^2.1.0",
     "react": "^19.1.0",
     "tsup": "^8.4.0",
     "vitest": "^3.1.1"
